Added "set_reference_state" wrapper for Mathcad and Updated Example Worksheets (#1056)
- Added Mathcad wrapper for set_reference_stateS(), designated
set_reference_state in Mathcad.
- Added to the funcdoc interface (CoolProp_EN.xml)
- Tidy up some formatting in the README files

* Updated Mathcad 15 and Prime Example Sheets + PDFs of Each

@ibell I know we don't like binaries in the rep, but I feel these are
critical.  They are updates to the Mathcad worksheets that were already
there and PDF versions for anyone who doesn't have Mathcad (i.e.
@ibell).  The files are only about 250 KB each or 1MB total.  I don't
expect we'll need to add to them or add more.  We might consider how to
put these out with the DLL binaries as well.  They are the only
documentation on how to use the wrappers correctly.

* Fixed CMakeList.txt confict and modified M15 worksheet from .xmcd to .xmcdz (compressed file type)