Change names of dlls copied to SMath wrapper: use underscore instead of dot
This is required in order to remove ".dll" extension from the library
name in .NET DllImport. And that, in turn, is intended to enable
cross-platform use of SMath wrapper on other OSes, like Linux (that is
supported by both CoolProp and SMath).
As explained at
http://www.mono-project.com/docs/advanced/pinvoke/, it's recommended to
omit the library extension to allow mono to find native libraries.