LibreOffice: Use pip for installing CoolProp python package
Updated the installation routine for the CoolProp python package in the
LibreOffice wrapper to use pip. This simplifies the script, as the download
links for different platforms and ABI versions will be handled by pip.

Bundled Python interpreters in LibreOffice doesn't include pip. But as pip
itself is platform independent and runs on py2/py3, it will be bundled inside
the LibreOffice extension at build time.