PhaseEnvelopeRoutines::is_inside above Pcr
Finds nearest point on envolope even if we are above the critical pressure. I've done this the lazy way of looping the entire envolope and for each point finding the closest point, however it would be more efficient to e.g. start at the critical pressure and go left or right.

Also I'm noting that the code for below critical pressure but gas or liquid uses the envelope point for the pressure and calls this the closest. I'm not sure if this is entirely correct - many phase envelopes are highly asymmetrical and the nearest point might not share the same pressure. My understanding is that the isotherms contour around the phase envelope so I’m assuming my approach works – but I’m probably making a horrible mistake here so a bit of feedback from someone with a deeper understanding of the physics would be helpful.
Also I’m doing this to fix the handling of supercritical mixtures – I.e. blends of room temperature gases raised to high pressure such as in a cylinder. And I’m not sure if this routine is called from anywhere using it differently. But because FlashRoutines::PT_flash_mixtures calls this and then with previous code deduces that the supercritical fluid is a liquid, it handles it as such. I'm not sure if PT_flash_mixtures should in fact have a different handling for superitical states. At present it can't handle anything except solutions right of the phase envolope.