Provide better feedback for bad DQ inputs (#2203)
* Provide better feedback for bad DQ inputs

DQ is poorly defined (not unique) for densities above critical
density. Provide the user better feedback when they try to do this,
but do not give them a non-unique answer that might give them
trouble later.

* Prevent DQ warning for saturation curves

The saturation curves (Q=0, Q=1) are monotonic and invertible, so we
should not tell the user otherwise.

* Add small tolerance to DQ quality checking

Exact checking of floating point values is bad practice, as
iterative calculations can introduce small errors. This commit
ensures that when the DQ flash routine checks for the saturation
curves that it allows for some slop.

* Prevent DQ evaluation when D>Dcrit

The combination of density and quality inputs are not valid when the
density is greater than the critical density and the quality is
greater than 0. A small tolerance is added to allow for numeric
slop.

* Add discussion of Q to DQ warning