Topic 2142 (#2223)
* Remove R1233zd(E) hL ancillary

The R1233zd(E) hL ancillary performs poorly. This can result in a
failure to converge when provided inputs (p/h) lie between the
ancillary saturation curve and the true saturation curve. Removing
the ancillary makes CoolProp fall back to a robust saturation
solver.

* Add 0.01 K cushion to Tmax in HSU_P_flash

Within HSU_P_flash, if an ancillary is used it can have a small
error. If the point of interest is near that saturation curve, this
error can prevent the solver from converging. Adding a 0.01 K margin
account for small ancillary errors and allow the solver to converge.

* Correct 0.01K cushion

The 0.01K adjustment should increase the potential range of the
solver. Previously, the minimum was increased, but it should be
decreased.