Fix Missing SWIG Wrapper
Added pass-through function to CoolProp.cpp for apply_simple_mixing_rule which was not recognised by SWIG and included in wrappers. Had to rename original function to apply_basic_mixing rule to avoid naming conflict. Using a similar approach as CoolProp::add_fluids_as_JSON and keeping the same function name failed to compile. Also added HumidAirProp.h to the SWIG_DEPENDENCIES so that these functions are included in the wrapper.