Addressed review comment
Reverted all previous changes to code as per request from code review. This version has no namespace changes and pass-through wrapping of missing function. Only change is to modify CoolProp.i for SWIG to ensure that the missing function, and that function only, is included when SWIG creates the wrapped code.