Fix Missing SWIG Wrapper (#2275)
* Fix Missing SWIG Wrapper

Added pass-through function to CoolProp.cpp for apply_simple_mixing_rule which was not recognised by SWIG and included in wrappers. Had to rename original function to apply_basic_mixing rule to avoid naming conflict. Using a similar approach as CoolProp::add_fluids_as_JSON and keeping the same function name failed to compile. Also added HumidAirProp.h to the SWIG_DEPENDENCIES so that these functions are included in the wrapper.

* Include SWIG wrapper without changing function names

Instead of changing function names, applied same approach as used with add_fluids_as_JSON to implement the passthrough function. apply_simple_mixing_rule is called in MixtureParameters.cpp using a namespace. This means that SWIG wraps the function and the underlying function names are not touched.

* Addressed review comment

Reverted all previous changes to code as per request from code review. This version has no namespace changes and pass-through wrapping of missing function. Only change is to modify CoolProp.i for SWIG to ensure that the missing function, and that function only, is included when SWIG creates the wrapped code.