Update REFPROPMixtureBackend.cpp
Coolprop library does not implement the function call setrefdll from Refprop. As a result, there are discrepancies between the calculation results of Refprop and Coolprop when computing predefined mixture MIX models, primarily in terms of standard state deviations (which affect enthalpy values and related quantities).

Coolprop does provide a method to set the standard state, which can be set to DEF. However, the default value of DEF in Coolprop still differs from the calculation results of Refprop.