Fix bug introduced by 443a2fd (#2351)
Commit 443a2fd addedsome extra checks for convergence in the Helmholtz
PHSU flash routines. However, its checking was more aggressive than the
original routine. This caused the routine to raise an exception on
several values that had previously been deemed acceptable. This trickled
down to cause the TTSE tables to be filled with empty values, making
them less robust.

This commit allows for a solution to be acceptable if the change is
small between iterations as was originally intended.